requirements.txt
streamlit
python-docx
Pillow
PyMuPDF
arabic-reshaper
python-bidi
